Bump minSdkVersion to 21, purge related lower-API workarounds
Fixes #7843
