+++ AnkiDroid/build.gradle
@@ -43,20 +43,18 @@ android {
         // needed for upgrades to be offered correctly.
         versionCode=21500111
         versionName="2.15alpha11"
-        minSdkVersion 16
+        minSdkVersion 21
         //noinspection OldTargetApi - also performed in api/build.fradle
         targetSdkVersion 29 // change .travis.yml platform download at same time
-        multiDexEnabled true
         testApplicationId "com.ichi2.anki.tests"
         vectorDrawables.useSupportLibrary = true
-        // This is a custom class, in androidx to override package-level functions
-        testInstrumentationRunner 'androidx.test.runner.MultiDexJUnitRunner'
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
 
         // This enables long timeouts required on slow ARM emulators, e.g. Travis
         adbOptions {
             //logLevel verbose // let's try to see what is happening
             timeOutInMs 20 * 60 * 1000  // 20 minutes
-            DdmPreferences.setTimeOut(120000) // apparently if not multidex you do this
+            DdmPreferences.setTimeOut(120000)
         }
     }
     signingConfigs {
@@ -224,7 +222,6 @@ dependencies {
     implementation 'androidx.browser:browser:1.3.0'
     implementation 'androidx.exifinterface:exifinterface:1.3.2'
     implementation 'androidx.localbroadcastmanager:localbroadcastmanager:1.0.0'
-    implementation "androidx.multidex:multidex:2.0.1"
     implementation "androidx.preference:preference:1.1.1"
     implementation 'androidx.recyclerview:recyclerview:1.1.0'
     implementation 'androidx.sqlite:sqlite-framework:2.1.0'
@@ -235,9 +232,8 @@ dependencies {
     implementation 'com.getbase:floatingactionbutton:1.10.1'
     // io.github.java-diff-utils:java-diff-utils is the natural successor here, but requires API24, #7091
     implementation 'org.bitbucket.cowwoc:diff-match-patch:1.2'
-    // #6419  - API 27 (& maybe others) could not perform new ZipFile() on a 2GB+ apkg
-    // noinspection GradleDependency - pinned at 1.12 until API21 minSdkVersion (File.toPath usage)
-    implementation 'org.apache.commons:commons-compress:1.12'
+    // noinspection GradleDependency - pinned at 1.12 until API26 minSdkVersion (File.toPath usage)
+    implementation 'org.apache.commons:commons-compress:1.12' // #6419 - handle >2GB apkg files
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
 
 
@@ -276,7 +272,6 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-contrib:3.3.0'
     androidTestImplementation 'androidx.test.ext:junit:1.1.2'
     androidTestImplementation 'androidx.test:rules:1.3.0'
-    androidTestImplementation 'org.smali:dexlib2:2.4.0'
 
     //For AnkiDroid JS API Versioning
     implementation "com.github.zafarkhaja:java-semver:0.9.0"
