+++ AnkiDroid/src/main/AndroidManifest.xml
@@ -43,8 +43,6 @@
                 android:description="@string/read_write_permission_description"
                 android:protectionLevel="dangerous"
                 />
-    <!-- custom tabs library has minimum SDK of 16 -->
-    <uses-sdk tools:overrideLibrary="android.support.customtabs, com.getbase.floatingactionbutton"/>
 
     <supports-screens
         android:anyDensity="true"
