+++ AnkiDroid/src/main/java/com/ichi2/anki/ReadText.java
@@ -19,18 +19,16 @@ package com.ichi2.anki;
 import android.content.Context;
 import android.content.res.Resources;
 import android.net.Uri;
+import android.os.Bundle;
 import android.os.Handler;
 import android.speech.tts.TextToSpeech;
 import android.speech.tts.UtteranceProgressListener;
 
-import android.view.View;
 import android.view.WindowManager;
 import android.widget.Toast;
 
 import com.afollestad.materialdialogs.MaterialDialog;
 import com.google.android.material.snackbar.Snackbar;
-import com.ichi2.compat.Compat;
-import com.ichi2.compat.CompatHelper;
 import com.ichi2.libanki.Sound;
 
 import java.lang.ref.WeakReference;
@@ -51,8 +49,7 @@ public class ReadText {
     private static int mOrd;
     private static Sound.SoundSide mQuestionAnswer;
     public static final String NO_TTS = "0";
-    private static final Compat compat = CompatHelper.getCompat();
-    private static final Object mTtsParams = compat.initTtsParams();
+    private static final Bundle mTtsParams = new Bundle();;
 
     public static Sound.SoundSide getmQuestionAnswer() {
         return mQuestionAnswer;
@@ -71,7 +68,7 @@ public class ReadText {
                 //sTextQueue.add(new String[] { text, loc });
             }
             Timber.d("tts text '%s' to be played for locale (%s)", text, loc);
-            compat.speak(mTts, mTextToSpeak, queueMode, mTtsParams, "stringId");
+            mTts.speak(mTextToSpeak, queueMode, mTtsParams, "stringId");
         }
     }
