--- AnkiDroid/src/main/java/com/ichi2/compat/CompatV18.java
@@ -1,30 +1,0 @@
-/***************************************************************************************
- * Copyright (c) 2018 Mike Hardy <github@mikehardy.net>                                 *
- *                                                                                      *
- * This program is free software; you can redistribute it and/or modify it under        *
- * the terms of the GNU General Public License as published by the Free Software        *
- * Foundation; either version 3 of the License, or (at your option) any later           *
- * version.                                                                             *
- *                                                                                      *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY      *
- * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A      *
- * PARTICULAR PURPOSE. See the GNU General Public License for more details.             *
- *                                                                                      *
- * You should have received a copy of the GNU General Public License along with         *
- * this program.  If not, see <http://www.gnu.org/licenses/>.                           *
- ****************************************************************************************/
-
-package com.ichi2.compat;
-
-import android.annotation.TargetApi;
-import android.os.StatFs;
-
-import androidx.annotation.CheckResult;
-
-@TargetApi(18)
-public class CompatV18 extends CompatV17 implements Compat {
-
-    @Override
-    @CheckResult
-    public long getAvailableBytes(StatFs stat) { return stat.getAvailableBytes(); }
-}
